Nests the user objects within UserResponse. Closes #24
